consumers/requirements.txt
kafka-python
cassandra-driver
sseclient-py
requests